FIX(FW5.STANDBY) 3C-STUCK STATE with a 10ms delay >
REASON: it stucks just before falling to sleep,
with state 3C: GPS_READY|MEMS_READY|NET_READY|CONNECTED,
flags are not WORKING anymore but not yet STANDBY,
so that `telemetry()` loop rolls in quick succession,
meaning that the CPU cannot task-switch(!??)
for `process()` to return, and STANDBY flag to become 1.

TODO: delaying the quick `telemetry()` loop is a lousy solution.